requirements.txt
streamlit==1.38.0
PyYAML==6.0.2
matplotlib==3.9.0
networkx==3.3
pandas==2.2.2
